onprem/voucher/requirements.txt
flask
uwsgi
bcrypt
pexpect
pyjwt
cryptography
requests
jose
